requirements.txt
prometheus_client==0.20.0
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0
requests==2.32.3
aiohttp==3.9.5
gunicorn==22.0.0
Flask==3.0.3
psycopg2-binary==2.9.9
pika==1.3.2
